additional refactor
