+++ Cargo.toml
@@ -32,7 +32,6 @@ serde_with = { version = "3.3.0", features = [] }
 thiserror = { version = "2.0.3" }
 uuid = { version = "1.3.3", features = ["serde", "v4"] }
 
-
 [package.metadata.git-cliff.changelog]
 header = "All notable changes to this project will be documented in this file."
 body = "..."
