+++ src/bin/load_news.rs
@@ -39,17 +39,17 @@ use alpha_vantage_rust::{
     alpha_io::news_loader::{load_news, Params},
     misc_functions::get_exe_name,
   },
-  db_funcs::{
-    get_proc_id_or_insert, log_proc_end, log_proc_start,
-  },
   dbfunctions::{
-    author::get_authors, base::establish_connection_or_exit, sources::get_sources,
+    author::get_authors,
+    base::establish_connection_or_exit,
+    combined::get_sids_and_names_with_overview,
+    process::{get_proc_id_or_insert, log_proc_end, log_proc_start},
+    sources::get_sources,
     topic_refs::get_topics,
   },
 };
 use dotenvy::dotenv;
 use indicatif::ProgressBar;
-use alpha_vantage_rust::dbfunctions::combined::get_sids_and_names_with_overview;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
   dotenv().ok();
