+++ src/bin/load_open_close.rs
@@ -37,14 +37,14 @@ use std::process;
 
 use alpha_vantage_rust::{
   alpha_lib::{alpha_io::base::load_summary, misc_functions::get_exe_name},
-  db_funcs::{
-    get_proc_id_or_insert, log_proc_end, log_proc_start,
+  dbfunctions::{
+    base::establish_connection_or_exit,
+    combined::get_sids_and_names_with_overview,
+    process::{get_proc_id_or_insert, log_proc_end, log_proc_start},
   },
-  dbfunctions::base::establish_connection_or_exit,
 };
 use dotenvy::dotenv;
 use indicatif::ProgressBar;
-use alpha_vantage_rust::dbfunctions::combined::get_sids_and_names_with_overview;
 
 fn main() {
   let conn = &mut establish_connection_or_exit();
