+++ src/lib.rs
@@ -2,9 +2,8 @@ extern crate core;
 
 pub mod alpha_lib;
 pub mod datatypes;
-pub mod db_funcs;
 pub mod db_models;
 pub mod dbfunctions;
-pub mod file_processors;
 pub mod schema;
 pub mod security_types;
+pub mod util;
