+++ src/util/processor.rs
@@ -26,81 +26,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 use std::{env, error::Error, fmt, fs::File, io::BufReader};
 
 use crate::datatypes::data_file_types::{DigitalAsset, NasdaqListed, NyseOtherSymbol};
-
-const NASDAQ: &str = "NASDAQ";
-const NYSE: &str = "NYSE";
-const MAX_SYMBOLS: usize = 10000;
-
-const DIGITAL: &str = "DIGITAL";
-
-#[derive(Debug, Clone)]
-struct UnknownExchangeError(String);
-
-impl fmt::Display for UnknownExchangeError {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    write!(f, "Unknown exchange type: {}", self.0)
-  }
-}
-
-impl Error for UnknownExchangeError {}
-
-/// Opens and reads an exchange CSV file, deserializing each record into either a `NasdaqListed` or
-/// `NyseOtherSymbol` struct based on the `exchange` parameter. Extracts the relevant symbol from
-/// each record and stores them in a vector.
-///
-/// # Arguments
-/// * `file_name` - A `String` that holds the name of the CSV file to be processed.
-/// * `exchange` - A string slice that denotes the type of exchange (NASDAQ or NYSE). This
-///   determines the type of struct that the CSV records are deserialized into.
-///
-/// # Returns
-/// * `Ok(Vec<String>)` - A vector of `String` objects, each representing the symbol from a record
-///   in the CSV file. The vector has an initial capacity of `MAX_SYMBOLS`.
-/// * `Err(Box<dyn Error>)` - A boxed dynamic Error object that might be returned in case of failure
-///   to open the file, failure to deserialize the records, or if an unknown exchange type is
-///   provided.
-///
-/// # Errors
-/// This function will return an `Err` variant if the file cannot be opened, if there is an error
-/// during deserialization of the CSV records, or if an unknown exchange type is passed.
-///
-/// It can also return a `UnknownExchangeError` if the `exchange` parameter does not match with
-/// either NASDAQ or NYSE.
-fn csv_proc(file_name: String, exchange: &str) -> Result<Vec<String>, Box<dyn Error>> {
-  let file = File::open(file_name)?;
-  let mut file_reader = csv::Reader::from_reader(BufReader::new(file));
-  let mut symbols: Vec<String> = Vec::with_capacity(MAX_SYMBOLS);
-
-  match exchange {
-    NASDAQ => {
-      for result in file_reader.deserialize() {
-        let record: NasdaqListed = result?;
-        symbols.push(record.symbol);
-      }
-    }
-    NYSE => {
-      for result in file_reader.deserialize() {
-        let record: NyseOtherSymbol = result?;
-        symbols.push(record.actsymbol);
-      }
-    }
-    DIGITAL => {
-      for result in file_reader.deserialize() {
-        let record: DigitalAsset = result?;
-        symbols.push(format!("{},{}", record.symbol, record.name));
-      }
-    }
-    _ => {
-      println!("csv_proc: unknown exchange type: {}", exchange);
-      return Err(Box::new(UnknownExchangeError(exchange.to_string())));
-    }
-  }
-  Ok(symbols)
-}
 
 /// Processes the root exchange data files and returns the list of securities.
 ///
@@ -147,3 +75,73 @@ pub fn file_proc(file_arr: Vec<(&str, &str)>) -> Result<Vec<Vec<String>>, Box<dy
 
   return Ok(security_vector);
 }
+
+/// Opens and reads an exchange CSV file, deserializing each record into either a `NasdaqListed` or
+/// `NyseOtherSymbol` struct based on the `exchange` parameter. Extracts the relevant symbol from
+/// each record and stores them in a vector.
+///
+/// # Arguments
+/// * `file_name` - A `String` that holds the name of the CSV file to be processed.
+/// * `exchange` - A string slice that denotes the type of exchange (NASDAQ or NYSE). This
+///   determines the type of struct that the CSV records are deserialized into.
+///
+/// # Returns
+/// * `Ok(Vec<String>)` - A vector of `String` objects, each representing the symbol from a record
+///   in the CSV file. The vector has an initial capacity of `MAX_SYMBOLS`.
+/// * `Err(Box<dyn Error>)` - A boxed dynamic Error object that might be returned in case of failure
+///   to open the file, failure to deserialize the records, or if an unknown exchange type is
+///   provided.
+///
+/// # Errors
+/// This function will return an `Err` variant if the file cannot be opened, if there is an error
+/// during deserialization of the CSV records, or if an unknown exchange type is passed.
+///
+/// It can also return a `UnknownExchangeError` if the `exchange` parameter does not match with
+/// either NASDAQ or NYSE.
+pub fn csv_proc(file_name: String, exchange: &str) -> Result<Vec<String>, Box<dyn Error>> {
+  let file = File::open(file_name)?;
+  let mut file_reader = csv::Reader::from_reader(BufReader::new(file));
+  let mut symbols: Vec<String> = Vec::with_capacity(MAX_SYMBOLS);
+
+  match exchange {
+    NASDAQ => {
+      for result in file_reader.deserialize() {
+        let record: NasdaqListed = result?;
+        symbols.push(record.symbol);
+      }
+    }
+    NYSE => {
+      for result in file_reader.deserialize() {
+        let record: NyseOtherSymbol = result?;
+        symbols.push(record.actsymbol);
+      }
+    }
+    DIGITAL => {
+      for result in file_reader.deserialize() {
+        let record: DigitalAsset = result?;
+        symbols.push(format!("{},{}", record.symbol, record.name));
+      }
+    }
+    _ => {
+      println!("csv_proc: unknown exchange type: {}", exchange);
+      return Err(Box::new(UnknownExchangeError(exchange.to_string())));
+    }
+  }
+  Ok(symbols)
+}
+
+pub const NASDAQ: &str = "NASDAQ";
+pub const NYSE: &str = "NYSE";
+pub const MAX_SYMBOLS: usize = 10000;
+pub const DIGITAL: &str = "DIGITAL";
+
+#[derive(Debug, Clone)]
+pub struct UnknownExchangeError(String);
+
+impl fmt::Display for UnknownExchangeError {
+  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    write!(f, "Unknown exchange type: {}", self.0)
+  }
+}
+
+impl Error for UnknownExchangeError {}
